requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.3.0
packaging==24.0
PySocks==1.7.1
python-dotenv==1.0.1
PyVirtualDisplay==3.0
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
